order-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
mysql-connector-python==9.2.0
pydantic==2.11.1
